refactor: isManualConfirmed 칼럼 삭제 및 관련 로직 분리
